make histories int16
passed multithreaded LTC:
Elo   | 3.70 +- 3.51 (95%)
SPRT  | 40.0+0.40s Threads=2 Hash=256MB
LLR   | 2.96 (-2.94, 2.94) [-3.50, 0.50]
Games | N: 10530 W: 2649 L: 2537 D: 5344
Penta | [57, 1206, 2613, 1346, 43]
https://nonlinear.eu.pythonanywhere.com/test/1494/
passed LTC:
Elo   | 1.87 +- 2.59 (95%)
SPRT  | 40.0+0.40s Threads=1 Hash=128MB
LLR   | 3.00 (-2.94, 2.94) [-3.50, 0.50]
Games | N: 17238 W: 4090 L: 3997 D: 9151
Penta | [41, 1936, 4571, 2031, 40]
https://nonlinear.eu.pythonanywhere.com/test/1481/
passed STC:
Elo   | 4.94 +- 3.90 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=32MB
LLR   | 3.13 (-2.94, 2.94) [-3.50, 0.50]
Games | N: 8864 W: 2327 L: 2201 D: 4336
Penta | [52, 1004, 2210, 1098, 68]
https://nonlinear.eu.pythonanywhere.com/test/1477/
bench: 2106376

+++ src/history.cpp
@@ -1,6 +1,6 @@
 #include "history.hpp"
 
-int& ContinuationHistory::get(Piece prev_piece, Square prev_to, Piece piece, Square to){
+int16_t& ContinuationHistory::get(Piece prev_piece, Square prev_to, Piece piece, Square to){
     return this->history[
         prev_piece * NUM_SQUARES * NUM_PIECES * NUM_SQUARES
       + prev_to.index() * NUM_PIECES * NUM_SQUARES
@@ -9,7 +9,7 @@ int& ContinuationHistory::get(Piece prev_piece, Square prev_to, Piece piece, Squ
     ];
 }
 
-int& FromToHistory::get(Color color, Square from, Square to){
+int16_t& FromToHistory::get(Color color, Square from, Square to){
     return this->history[
         color * NUM_SQUARES * NUM_SQUARES
       + from.index() * NUM_SQUARES
@@ -17,7 +17,7 @@ int& FromToHistory::get(Color color, Square from, Square to){
     ];
 }
 
-int& CaptureHistory::get(Piece piece, Square to, Piece captured){
+int16_t& CaptureHistory::get(Piece piece, Square to, Piece captured){
     // if the move was en passant, this function may be called with captured == None,
     // but idx will still be less than the history size
     int idx = piece * NUM_SQUARES * NUM_PIECETYPES 
@@ -28,18 +28,18 @@ int& CaptureHistory::get(Piece piece, Square to, Piece captured){
     return history[idx];
 }
 
-int& PawnCorrectionHistory::get(Color color, uint16_t key){
+int16_t& PawnCorrectionHistory::get(Color color, uint16_t key){
     return this->history[NUM_COLORS * (key % PAWN_CORRHIST_SIZE) + color];
 }
 
-int& MinorCorrectionHistory::get(Color color, uint16_t key){
+int16_t& MinorCorrectionHistory::get(Color color, uint16_t key){
     return this->history[NUM_COLORS * (key % MINOR_CORRHIST_SIZE) + color];
 }
 
-int& MajorCorrectionHistory::get(Color color, uint16_t key){
+int16_t& MajorCorrectionHistory::get(Color color, uint16_t key){
     return this->history[NUM_COLORS * (key % MAJOR_CORRHIST_SIZE) + color];
 }
 
-int& NonPawnCorrectionHistory::get(Color color, uint16_t key){
+int16_t& NonPawnCorrectionHistory::get(Color color, uint16_t key){
     return this->history[NUM_COLORS * (key % NONPAWN_CORRHIST_SIZE) + color];
 }
